♻️Refactor: 공고 리스트, 상세 공고 페이지 폴더 이동
