+++ components/common/Header.tsx
@@ -107,7 +107,7 @@ const Header = () => {
     e.preventDefault();
     if (searchKeyword.trim()) {
       router.push(
-        `/auth/notices?search=${encodeURIComponent(searchKeyword.trim())}`
+        `/staff/notices?search=${encodeURIComponent(searchKeyword.trim())}`
       );
     }
   };
